Extract contributors of the project from both sidebars.
