+++ src/components/projects/project/project.module.scss
@@ -35,48 +35,6 @@
   line-height: 1.4;
 }
 
-.contributors {
-  width: 350px;
-  background-color: #141414;
-  padding: 1rem;
-  margin-top: 1rem;
-  margin-bottom: 1rem;
-
-  h4 {
-    margin-bottom: 12px;
-  }
-}
-
-.avatars {
-  position: sticky;
-  display: flex;
-  flex-wrap: wrap;
-  align-items: center;
-  width: 100%;
-  margin-left: 5px;
-
-  li {
-    position: sticky;
-    width: 30px;
-    height: 30px;
-    margin-left: -5px;
-    border: 1px solid gray;
-    border-radius: 50px;
-
-    &:last-child {
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      color: #f4f4f4;
-      cursor: pointer;
-    }
-  }
-
-  img {
-    border-radius: 50px;
-  }
-}
-
 @media screen and (max-width: 680px) {
   .wrapper {
     display: block;
@@ -99,10 +57,5 @@
     font-size: 1rem;
     margin-right: 0;
     margin-bottom: 0;
-  }
-
-  .contributors {
-    width: 100%;
-    padding: 0;
   }
 }
